chore(hogql): rename HogQLPrinter to BasePrinter, add dedicated HogQLPrinter
Preparatory refactor for splitting dialect-specific printer logic out of
the shared base. The existing HogQLPrinter class served three roles —
shared AST visitor, HogQL output printer, and parent for the CH/PG
subclasses. This splits it so each dialect has its own dedicated class
inheriting from a neutral BasePrinter.

No behavior change. Subsequent PRs will migrate dialect branches from
BasePrinter into their respective subclasses.
