+++ posthog/hogql/printer/__init__.py
@@ -1,5 +1,6 @@
-from posthog.hogql.printer.base import HogQLPrinter
+from posthog.hogql.printer.base import BasePrinter
 from posthog.hogql.printer.clickhouse import ClickHousePrinter
+from posthog.hogql.printer.hogql import HogQLPrinter
 from posthog.hogql.printer.postgres import PostgresPrinter
 from posthog.hogql.printer.utils import (
     prepare_and_print_ast,
@@ -13,6 +14,7 @@ __all__ = [
     "prepare_ast_for_printing",
     "print_prepared_ast",
     "to_printed_hogql",
+    "BasePrinter",
     "HogQLPrinter",
     "ClickHousePrinter",
     "PostgresPrinter",
